feat: Added leave feature, test, code clean up and code improvements

+++ apps/backend/src/main/kotlin/navikt/appsec/securitychampionstats/stats/AdminController.kt
@@ -4,8 +4,11 @@ import navikt.appsec.securitychampionstats.common.hikari.PostgresRepository
 import navikt.appsec.securitychampionstats.stats.dto.DeleteMember
 import navikt.appsec.securitychampionstats.stats.dto.MemberInfo
 import navikt.appsec.securitychampionstats.stats.dto.Points
+import navikt.appsec.securitychampionstats.utils.Validate
+import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -18,9 +21,17 @@ import java.util.UUID
 class AdminController(
     private val repo: PostgresRepository,
 ) {
-
+    private val logger = LoggerFactory.getLogger(AdminController::class.java)
+    private val validate = Validate()
     @PostMapping("/member")
     fun addMember(@RequestBody memberInfo: MemberInfo): ResponseEntity<Any>{
+        logger.info("Request to add member by user" +
+                " ${SecurityContextHolder.getContext().authentication.name}")
+        if (!validate.isValidEmail(memberInfo.email) or !validate.isValidName(memberInfo.fullname)) {
+            logger.warn("Attempt to add member failed due to invalid email format, " +
+                    "request made by user ${SecurityContextHolder.getContext().authentication.name}")
+            return ResponseEntity.status(HttpStatus.ACCEPTED).build()
+        }
         val id = UUID.randomUUID().toString()
         repo.addMember(memberInfo.fullname, id = id, memberInfo.email)
         return ResponseEntity("User was created", HttpStatus.CREATED)
@@ -28,18 +39,28 @@ class AdminController(
 
     @DeleteMapping("/member")
     fun deleteMember(@RequestBody member: DeleteMember): ResponseEntity<Any>{
+        logger.info("Request to delete member was made by user " +
+                "${SecurityContextHolder.getContext().authentication.name}")
+        if (!validate.isValidEmail(member.email)) {
+            logger.warn("Attempt to delete member failed due to invalid email format, " +
+                    "request made by user ${SecurityContextHolder.getContext().authentication.name}")
+            return ResponseEntity.status(HttpStatus.ACCEPTED).build()
+        }
         repo.deleteMember(member.email)
         return ResponseEntity.status(HttpStatus.ACCEPTED).build()
     }
 
     @PostMapping("/points")
     fun addPoints(@RequestBody points: Points): ResponseEntity<Any>{
-        val result = repo.addPoints(points.email, points.points)
-        return if (result == 1) {
-            ResponseEntity.status(HttpStatus.ACCEPTED).build()
-        } else {
-            ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build()
+        logger.info("Request to add points was made by user " +
+                "${SecurityContextHolder.getContext().authentication.name}")
+        if (!validate.isValidEmail(points.email) or !validate.isValidNumber(points.points.toString())) {
+            logger.warn("Attempt to add points for user failed due to invalid email format, " +
+                    "request made by user ${SecurityContextHolder.getContext().authentication.name}")
+            return ResponseEntity.status(HttpStatus.ACCEPTED).build()
         }
+        repo.addPoints(points.email, points.points)
+        return ResponseEntity.status(HttpStatus.ACCEPTED).build()
     }
 
 }
